Move remaining s2s part to parachain
