+++ frame/bridge/message-gadget/Cargo.toml
@@ -12,7 +12,6 @@ version     = "2.9.4"
 [dependencies]
 # crates.io
 codec      = { package = "parity-scale-codec", version = "2.3", default-features = false }
-ethabi     = { version = "15.0", default-features = false }
 scale-info = { version = "1.0", default-features = false, features = ["derive"] }
 sha3       = { version = "0.9", default-features = false }
 # darwinia-network
@@ -21,7 +20,6 @@ darwinia-ethereum = { default-features = false, path = "../../dvm/ethereum" }
 frame-support = { default-features = false, git = "https://github.com/darwinia-network/substrate", branch = "darwinia-v0.12.4" }
 frame-system  = { default-features = false, git = "https://github.com/darwinia-network/substrate", branch = "darwinia-v0.12.4" }
 sp-core       = { default-features = false, git = "https://github.com/darwinia-network/substrate", branch = "darwinia-v0.12.4" }
-sp-std        = { default-features = false, git = "https://github.com/darwinia-network/substrate", branch = "darwinia-v0.12.4" }
 
 [features]
 default = ["std"]
@@ -29,7 +27,6 @@ default = ["std"]
 std = [
 	# crates.io
 	"codec/std",
-	"ethabi/std",
 	"scale-info/std",
 	"sha3/std",
 	# darwinia-network
