+++ frame/support/src/lib.rs
@@ -28,41 +28,3 @@ pub mod traits;
 pub mod balance {
 	pub use crate::structs::{StakingLock, Unbonding};
 }
-
-// TODO: Should we move this to `s2s-primitives`?
-pub mod s2s {
-	// --- crates.io ---
-	use codec::Encode;
-	// --- darwinia-network ---
-	use ethereum_primitives::H256;
-	// --- paritytech ---
-	use bp_messages::{LaneId, MessageNonce};
-	use bp_runtime::{derive_account_id, ChainId, SourceAccount};
-	use frame_support::ensure;
-	use sp_runtime::{
-		traits::{BadOrigin, Convert},
-		DispatchError,
-	};
-	use sp_std::cmp::PartialEq;
-
-	pub trait LatestMessageNoncer {
-		fn outbound_latest_generated_nonce(lane_id: LaneId) -> MessageNonce;
-		fn inbound_latest_received_nonce(lane_id: LaneId) -> MessageNonce;
-	}
-
-	pub fn ensure_source_account<AccountId, Converter>(
-		chain_id: ChainId,
-		source_account: AccountId,
-		derived_account: &AccountId,
-	) -> Result<(), DispatchError>
-	where
-		AccountId: PartialEq + Encode,
-		Converter: Convert<H256, AccountId>,
-	{
-		let hex_id =
-			derive_account_id::<AccountId>(chain_id, SourceAccount::Account(source_account));
-		let target_id = Converter::convert(hex_id);
-		ensure!(&target_id == derived_account, BadOrigin);
-		Ok(())
-	}
-}
