+++ node/runtime/pangolin/src/pallets/to_parachain_backing_.rs
@@ -8,8 +8,7 @@ use bp_message_dispatch::CallOrigin;
 use bp_messages::LaneId;
 use bp_runtime::{messages::DispatchFeePayment, ChainId, PANGOLIN_PARACHAIN_CHAIN_ID};
 use bridge_runtime_common::lanes::PANGOLIN_PANGOLIN_PARACHAIN_LANE;
-use darwinia_support::s2s::LatestMessageNoncer;
-use to_parachain_backing::{Config, IssueFromRemotePayload, IssuingCall};
+use to_parachain_backing::{Config, IssueFromRemotePayload, IssuingCall, LatestMessageNoncer};
 
 pub struct PangolinParachainMessageNoncer;
 impl LatestMessageNoncer for PangolinParachainMessageNoncer {
